SPecONet/requirements.txt
pytorch 1.9.0
torchvision==0.10.0
torchaudio
pytorch-cuda=11.6
numpy==1.18.1
tqdm==4.42.1
pandas==1.0.1
matplotlib==3.1.3
scipy 1.6.2
seaborn 0.11.2


conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch

import pandas as pd
print(pd.__version__)

**Check gpu
import torch
>>> torch.cuda.is_available()
True

>>> torch.cuda.device_count()
1

>>> torch.cuda.current_device()
0

>>> torch.cuda.device(0)
<torch.cuda.device at 0x7efce0b03be0>

>>> torch.cuda.get_device_name(0)
'GeForce GTX 950M'